+++ tests/parser-functions.ts
@@ -13,7 +13,7 @@ export default {
     await page.waitFor('h1');
   },
 
-  async beforeHackerRankProblem(page: Page) {
+  async beforeHackerRank(page: Page) {
     await page.waitFor('.problem-statement');
   },
 
@@ -21,14 +21,18 @@ export default {
     await page.waitFor('.ant-card-body pre');
   },
 
-  async beforeOldGoogleCodeJam(page: Page) {
-    await page.waitFor('#dsb-problem-title0');
-  },
-
   async beforeNewGoogleCodeJam(page: Page) {
     await page.waitFor('.adventures > tr:last-child > td:nth-child(4) > a');
     await page.click('.adventures > tr:last-child > td:nth-child(4) > a');
     await page.waitFor('.collection > a:nth-child(2)');
     await page.click('.collection > a:nth-child(2)');
+  },
+
+  async beforeOldGoogleCodeJam(page: Page) {
+    await page.waitFor('#dsb-problem-title0');
+  },
+
+  async beforeQDUOJ(page: Page) {
+    await page.waitFor('#problem-main');
   },
 };