+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 2.3.0 - September 8th, 2018
 - Fixed bugs in the HackerRank, URI Online Judge and Codeforces parsers
-- Added parsers for UVa Online Judge, ICPC Live Archive, FZU Online Judge, CSU-ACM Online Judge, Hrbust Online Judge, HIT Online Judge and hihoCoder
+- Added parsers for UVa Online Judge, ICPC Live Archive, FZU Online Judge, CSU-ACM Online Judge, Hrbust Online Judge, HIT Online Judge, hihoCoder, ECNU Online Judge, SUSTech Online Judge, QDUOJ and NYTD Online Judge
 
 # 2.2.4 - June 30th, 2018
 - Added support for AtCoder Beta